requirements.txt
pinecone-client
langchain
openai
slack_sdk